move Wikidata and target data gathering logic. Closes #128
