+++ soweego/validator/checks.py
@@ -11,22 +11,24 @@ __copyright__ = 'Copyleft 2018, Hjfocs'
 
 import json
 import logging
-import re
 from collections import defaultdict
 
 import click
-import regex
-from sqlalchemy import or_
 
-from soweego.commons import url_utils
-from soweego.commons.cache import cached
 from soweego.commons.constants import HANDLED_ENTITIES, TARGET_CATALOGS
-from soweego.commons.db_manager import DBManager
+from soweego.commons.data_gathering import (connect_to_db,
+                                            extract_ids_from_urls,
+                                            gather_identifiers,
+                                            gather_relevant_pids,
+                                            gather_target_links,
+                                            gather_target_metadata,
+                                            gather_wikidata_links,
+                                            gather_wikidata_metadata)
 from soweego.importer.models.base_entity import BaseEntity
 from soweego.importer.models.musicbrainz_entity import (MusicbrainzArtistEntity,
                                                         MusicbrainzBandEntity)
 from soweego.ingestor import wikidata_bot
-from soweego.wikidata import api_requests, sparql_queries, vocabulary
+from soweego.wikidata import sparql_queries, vocabulary
 
 LOGGER = logging.getLogger(__name__)
 
@@ -57,7 +59,7 @@ def check_existence_cli(wikidata_query, class_qid, catalog_pid, database_table, 
 
 def check_existence(class_or_occupation_query, class_qid, catalog_pid, entity: BaseEntity):
     query_type = 'identifier', class_or_occupation_query
-    session = _connect_to_db()
+    session = connect_to_db()
     invalid = defaultdict(set)
     count = 0
 
@@ -74,12 +76,6 @@ def check_existence(class_or_occupation_query, class_qid, catalog_pid, entity: B
     LOGGER.info('Total invalid identifiers = %d', count)
     # Sets are not serializable to JSON, so cast them to lists
     return {target_id: list(qids) for target_id, qids in invalid.items()}
-
-
-def _connect_to_db():
-    db_manager = DBManager()
-    session = db_manager.new_session()
-    return session
 
 
 @click.command()
@@ -134,7 +130,7 @@ def check_links(entity, catalog, wikidata_cache=None):
     catalog_terms = _get_vocabulary(catalog)
 
     # Target links
-    target = _gather_target_links(entity, catalog)
+    target = gather_target_links(entity, catalog)
     # Early stop in case of no target links
     if target is None:
         return None, None, None
@@ -146,9 +142,9 @@ def check_links(entity, catalog, wikidata_cache=None):
         wikidata = {}
 
         # Wikidata links
-        _gather_identifiers(entity, catalog, catalog_terms['pid'], wikidata)
-        url_pids, ext_id_pids_to_urls = _gather_relevant_pids()
-        _gather_wikidata_links(wikidata, url_pids, ext_id_pids_to_urls)
+        gather_identifiers(entity, catalog, catalog_terms['pid'], wikidata)
+        url_pids, ext_id_pids_to_urls = gather_relevant_pids()
+        gather_wikidata_links(wikidata, url_pids, ext_id_pids_to_urls)
     else:
         wikidata = wikidata_cache
 
@@ -156,7 +152,7 @@ def check_links(entity, catalog, wikidata_cache=None):
     _assess('links', wikidata, target, to_deprecate, to_add)
 
     # Separate external IDs from URLs
-    ext_ids_to_add, urls_to_add = _extract_ids_from_urls(
+    ext_ids_to_add, urls_to_add = extract_ids_from_urls(
         to_add, ext_id_pids_to_urls)
 
     LOGGER.info('Validation completed. %d %s IDs to be deprecated, %d external IDs to be added, %d URL statements to be added', len(
@@ -215,7 +211,7 @@ def check_metadata(entity, catalog, wikidata_cache=None):
     catalog_terms = _get_vocabulary(catalog)
 
     # Target metadata
-    target = _gather_target_metadata(entity, catalog)
+    target = gather_target_metadata(entity, catalog)
     # Early stop in case of no target metadata
     if target is None:
         return None, None, None
@@ -227,8 +223,8 @@ def check_metadata(entity, catalog, wikidata_cache=None):
         wikidata = {}
 
         # Wikidata metadata
-        _gather_identifiers(entity, catalog, catalog_terms['pid'], wikidata)
-        _gather_wikidata_metadata(wikidata)
+        gather_identifiers(entity, catalog, catalog_terms['pid'], wikidata)
+        gather_wikidata_metadata(wikidata)
     else:
         wikidata = wikidata_cache
 
@@ -259,142 +255,6 @@ def _load_wikidata_cache(file_handle):
                 else:
                     cache[qid] = {data_type: set(value_list)}
     return cache
-
-
-@cached
-def _gather_target_metadata(entity_type, catalog):
-    catalog_constants = _get_catalog_constants(catalog)
-    catalog_entity = _get_catalog_entity(entity_type, catalog_constants)
-
-    LOGGER.info(
-        'Gathering %s birth/death dates/places and gender metadata ...', catalog)
-    entity = catalog_entity['entity']
-    # Base metadata
-    query_fields = _build_metadata_query_fields(entity, entity_type, catalog)
-
-    session = _connect_to_db()
-    result = None
-    try:
-        result = _run_metadata_query(
-            session, query_fields, entity, catalog, entity_type)
-        session.commit()
-    except:
-        session.rollback()
-        raise
-    finally:
-        session.close()
-
-    if not result:
-        return None
-    return _parse_target_metadata_query_result(result)
-
-
-def _run_metadata_query(session, query_fields, entity, catalog, entity_type):
-    query = session.query(
-        *query_fields).filter(or_(entity.born.isnot(None), entity.died.isnot(None)))
-    count = query.count()
-    if count == 0:
-        LOGGER.warning(
-            "No metadata available for %s %s. Stopping validation here", catalog, entity_type)
-        return None
-    LOGGER.info('Got %d entries with metadata from %s %s',
-                count, catalog, entity_type)
-    result_set = query.all()
-    return result_set
-
-
-def _build_metadata_query_fields(entity, entity_type, catalog):
-    # Base metadata
-    query_fields = [entity.catalog_id, entity.born,
-                    entity.born_precision, entity.died, entity.died_precision]
-    # Check additional metadata
-    if hasattr(entity, 'gender'):
-        query_fields.append(entity.gender)
-    else:
-        LOGGER.info('%s %s has no gender information', catalog, entity_type)
-    if hasattr(entity, 'birth_place'):
-        query_fields.append(entity.birth_place)
-    else:
-        LOGGER.info('%s %s has no birth place information',
-                    catalog, entity_type)
-    if hasattr(entity, 'death_place'):
-        query_fields.append(entity.death_place)
-    else:
-        LOGGER.info('%s %s has no death place information',
-                    catalog, entity_type)
-    return query_fields
-
-
-# Default date precision when not available: 9 (year)
-def _parse_target_metadata_query_result(result_set):
-    for result in result_set:
-        identifier = result.catalog_id
-        born = result.born
-        if born:
-            born_precision = getattr(result, 'born_precision', 9)
-            date_of_birth = f'{born}/{born_precision}'
-            yield identifier, vocabulary.DATE_OF_BIRTH, date_of_birth
-        else:
-            LOGGER.debug('%s: no birth date available', identifier)
-        died = result.died
-        if died:
-            died_precision = getattr(result, 'died_precision', 9)
-            date_of_death = f'{died}/{died_precision}'
-            yield identifier, vocabulary.DATE_OF_DEATH, date_of_death
-        else:
-            LOGGER.debug('%s: no death date available', identifier)
-        if hasattr(result, 'gender'):
-            yield identifier, vocabulary.SEX_OR_GENDER, result.gender
-        else:
-            LOGGER.debug('%s: no gender available', identifier)
-        if hasattr(result, 'birth_place'):
-            yield identifier, vocabulary.PLACE_OF_BIRTH, result.birth_place
-        else:
-            LOGGER.debug('%s: no birth place available', identifier)
-        if hasattr(result, 'death_place'):
-            yield identifier, vocabulary.PLACE_OF_DEATH, result.death_place
-        else:
-            LOGGER.debug('%s: no death place available', identifier)
-
-
-@cached
-def _gather_target_links(entity_type, catalog):
-    catalog_constants = _get_catalog_constants(catalog)
-    catalog_entity = _get_catalog_entity(entity_type, catalog_constants)
-
-    LOGGER.info('Gathering %s %s links ...', catalog, entity_type)
-    link_entity = catalog_entity['link_entity']
-
-    session = _connect_to_db()
-    result = None
-    try:
-        query = session.query(link_entity.catalog_id, link_entity.url)
-        count = query.count()
-        if count == 0:
-            LOGGER.warning(
-                "No links available for %s %s. Stopping validation here", catalog, entity_type)
-            return None
-        LOGGER.info('Got %d links from %s %s', count, catalog, entity_type)
-        result = query.all()
-        session.commit()
-    except:
-        session.rollback()
-        raise
-    finally:
-        session.close()
-
-    if result is None:
-        return None
-    for row in result:
-        yield row.catalog_id, row.url
-
-
-def _get_catalog_entity(entity, catalog_constants):
-    catalog_entity = catalog_constants.get(entity)
-    if not catalog_entity:
-        raise ValueError('Bad entity type: %s. Please use one of %s' %
-                         (entity, catalog_constants.keys()))
-    return catalog_entity
 
 
 def _assess(criterion, source, target_iterator, to_deprecate, to_add):
@@ -445,110 +305,12 @@ def _consume_target_iterator(target_iterator):
     return target
 
 
-def _extract_ids_from_urls(to_add, ext_id_pids_to_urls):
-    LOGGER.info('Starting extraction of IDs from target links to be added ...')
-    ext_ids_to_add = []
-    urls_to_add = []
-    for qid, urls in to_add.items():
-        for url in urls:
-            ext_id, pid = url_utils.get_external_id_from_url(
-                url, ext_id_pids_to_urls)
-            if ext_id:
-                ext_ids_to_add.append((qid, pid, ext_id))
-            else:
-                urls_to_add.append(
-                    (qid, vocabulary.DESCRIBED_AT_URL, url))
-    return ext_ids_to_add, urls_to_add
-
-
 def _get_vocabulary(catalog):
     catalog_terms = vocabulary.CATALOG_MAPPING.get(catalog)
     if not catalog_terms:
         raise ValueError('Bad catalog: %s. Please use one of %s' %
                          (catalog, vocabulary.CATALOG_MAPPING.keys()))
     return catalog_terms
-
-
-def _get_catalog_constants(catalog):
-    catalog_constants = TARGET_CATALOGS.get(catalog)
-    if not catalog_constants:
-        raise ValueError('Bad catalog: %s. Please use on of %s' %
-                         (catalog, TARGET_CATALOGS.keys()))
-    return catalog_constants
-
-
-def _gather_wikidata_metadata(wikidata):
-    LOGGER.info(
-        'Gathering Wikidata birth/death dates/places and gender metadata. This will take a while ...')
-    total = 0
-    # Generator of generators
-    for entity in api_requests.get_metadata(wikidata.keys()):
-        for qid, pid, value in entity:
-            parsed = _parse_wikidata_metadata_value(value)
-            if not wikidata[qid].get('metadata'):
-                wikidata[qid]['metadata'] = set()
-            wikidata[qid]['metadata'].add((pid, parsed))
-            total += 1
-    LOGGER.info('Got %d statements', total)
-
-
-def _parse_wikidata_metadata_value(value):
-    # Values: birth/death DATES, gender, birth/death places QIDs
-    date_value = value.get('time')
-    if date_value:
-        # +1180-01-01T00:00:00Z -> 1180-01-01
-        parsed = date_value[1:].split('T')[0]
-        return f"{parsed}/{value['precision']}"  # Date
-    item_value = value.get('id')
-    if item_value:
-        return item_value  # QID
-    return value  # String
-
-
-def _gather_wikidata_links(wikidata, url_pids, ext_id_pids_to_urls):
-    LOGGER.info(
-        'Gathering Wikidata sitelinks, third-party links, and external identifier links. This will take a while ...')
-    total = 0
-    for iterator in api_requests.get_links(wikidata.keys(), url_pids, ext_id_pids_to_urls):
-        for qid, url in iterator:
-            if not wikidata[qid].get('links'):
-                wikidata[qid]['links'] = set()
-            wikidata[qid]['links'].add(url)
-            total += 1
-    LOGGER.info('Got %d links', total)
-
-
-def _gather_relevant_pids():
-    url_pids = set()
-    for result in sparql_queries.url_pids_query():
-        url_pids.add(result)
-    ext_id_pids_to_urls = defaultdict(dict)
-    for result in sparql_queries.external_id_pids_and_urls_query():
-        for pid, formatters in result.items():
-            for formatter_url, formatter_regex in formatters.items():
-                if formatter_regex:
-                    try:
-                        compiled_regex = re.compile(formatter_regex)
-                    except re.error:
-                        LOGGER.debug(
-                            "Using 'regex' third-party library. Formatter regex not supported by the 're' standard library: %s", formatter_regex)
-                        compiled_regex = regex.compile(formatter_regex)
-                else:
-                    compiled_regex = None
-                ext_id_pids_to_urls[pid][formatter_url] = compiled_regex
-    return url_pids, ext_id_pids_to_urls
-
-
-def _gather_identifiers(entity, catalog, catalog_pid, aggregated):
-    catalog_constants = _get_catalog_constants(catalog)
-    LOGGER.info('Gathering Wikidata items with %s identifiers ...', catalog)
-    query_type = 'identifier', HANDLED_ENTITIES.get(entity)
-    for result in sparql_queries.run_identifier_or_links_query(query_type, catalog_constants[entity]['qid'], catalog_pid, 0):
-        for qid, target_id in result.items():
-            if not aggregated.get(qid):
-                aggregated[qid] = {'identifiers': set()}
-            aggregated[qid]['identifiers'].add(target_id)
-    LOGGER.info('Got %d %s identifiers', len(aggregated), catalog)
 
 
 def _upload_links(catalog, to_deprecate, urls_to_add, ext_ids_to_add, sandbox):
